Now calculates horizontal moves for switch
getHorizontalIndex is sent a rotated tetromino as a parameter, and it returns the initial horizontal index (which is how many moves right from the furthest left on the board it can be). With this index, we can calculate horizontal moves, which is how right move inputs to send to the Switch (if 0, none, if negative, that many lefts).